+++ src/tui_executor/tasks.py
@@ -1,13 +1,31 @@
 __all__ = [
     "TaskButton",
-    "ButtonPanel",
     "TaskKind",
 ]
 
 from enum import IntEnum
+from typing import Callable
 
 from textual.widgets import Button
-from textual.widgets import Collapsible
+
+
+def get_task_button_label(func: Callable, label: str = None) -> str:
+    """
+    Determine the label for the task button associated to the given function. If the task was
+    given a display name, that will have precedence over the given label. If both are None,
+    the function name will be returned.
+
+    Args:
+        func: the function that is associated with the button
+        label: an optional label that will have precedence over the function name as a label
+
+    Returns:
+        A string representing the label for the button.
+    """
+    if hasattr(func, "__ui_display_name__") and func.__ui_display_name__:
+        return func.__ui_display_name__
+    else:
+        return label or func.__name__
 
 
 class TaskKind(IntEnum):
@@ -18,24 +36,18 @@ class TaskKind(IntEnum):
 
 
 class TaskButton(Button):
-    def __init__(self, name: str):
+    """
+    A button that represents a task. Pressing the button will open its arguments panel.
+    """
+
+    DEFAULT_CSS = """
+    TaskButton {
+        min-width: 80%;
+    }
+    """
+
+    def __init__(self, name: str, function: Callable):
         super().__init__(name=name)
-        self._name = name
 
-
-class ButtonPanel(Collapsible):
-    """
-    A ButtonPanel is a collapsible widget that contains task buttons. The task buttons are generated from
-    the decorated functions in a Python module. One Python module (`*.py` file) will correspond to one ButtonPanel.
-    The functions are decorated with @exec_task (or the deprecated @exec_ui).
-
-    The title of the collapsible widget is the module name unless the variable UI_MODULE_DISPLAY_NAME was defined
-    in the module.
-    """
-    def __init__(self, title: str):
-        super().__init__(title=title)
-
-        self.task_buttons = []
-
-    def add_button(self, button: TaskButton):
-        self.task_buttons.append(button)
+        self.label = get_task_button_label(function)  # noqa
+        self.function = function