+++ tests/tasks/shared/unit_tests/module_not_found_global.py
@@ -1,4 +1,7 @@
 
+# The following statement will generate a ModuleNotFoundError during import.
+# This will be when the task is loaded at startup of the tui-executor.
+
 import a_non_existing_module